Create a child for my Calculator.js
